// src/grpc/proto/plugin/grpc_broker.ts
export interface ConnInfo {
  serviceId: number
  network: string
  address: string
  knock: ConnInfo_Knock | undefined
}

export interface ConnInfo_Knock {
  knock: boolean
  ack: boolean
  error: string
}

function createBaseConnInfo(): ConnInfo {
  return { serviceId: 0, network: '', address: '', knock: undefined }
}

export const ConnInfo = {
  encode(message: ConnInfo, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.serviceId !== 0) {
      writer.uint32(8).uint32(message.serviceId)
    }
    if (message.network !== '') {
      writer.uint32(18).string(message.network)
    }
    if (message.address !== '') {
      writer.uint32(26).string(message.address)
    }
    if (message.knock !== undefined) {
      ConnInfo_Knock.encode(message.knock, writer.uint32(34).fork()).ldelim()
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): ConnInfo {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseConnInfo()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break
          }

          message.serviceId = reader.uint32()
          continue
        case 2:
          if (tag !== 18) {
            break
          }

          message.network = reader.string()
          continue
        case 3:
          if (tag !== 26) {
            break
          }

          message.address = reader.string()
          continue
        case 4:
          if (tag !== 34) {
            break
          }

          message.knock = ConnInfo_Knock.decode(reader, reader.uint32())
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): ConnInfo {
    return {
      serviceId: isSet(object.serviceId) ? globalThis.Number(object.serviceId) : 0,
      network: isSet(object.network) ? globalThis.String(object.network) : '',
      address: isSet(object.address) ? globalThis.String(object.address) : '',
      knock: isSet(object.knock) ? ConnInfo_Knock.fromJSON(object.knock) : undefined,
    }
  },

  toJSON(message: ConnInfo): unknown {
    const obj: any = {}
    if (message.serviceId !== 0) {
      obj.serviceId = Math.round(message.serviceId)
    }
    if (message.network !== '') {
      obj.network = message.network
    }
    if (message.address !== '') {
      obj.address = message.address
    }
    if (message.knock !== undefined) {
      obj.knock = ConnInfo_Knock.toJSON(message.knock)
    }
    return obj
  },

  create<I extends Exact<DeepPartial<ConnInfo>, I>>(base?: I): ConnInfo {
    return ConnInfo.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<ConnInfo>, I>>(object: I): ConnInfo {
    const message = createBaseConnInfo()
    message.serviceId = object.serviceId ?? 0
    message.network = object.network ?? ''
    message.address = object.address ?? ''
    message.knock =
      object.knock !== undefined && object.knock !== null
        ? ConnInfo_Knock.fromPartial(object.knock)
        : undefined
    return message
  },
}

function createBaseConnInfo_Knock(): ConnInfo_Knock {
  return { knock: false, ack: false, error: '' }
}

export const ConnInfo_Knock = {
  encode(message: ConnInfo_Knock, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.knock !== false) {
      writer.uint32(8).bool(message.knock)
    }
    if (message.ack !== false) {
      writer.uint32(16).bool(message.ack)
    }
    if (message.error !== '') {
      writer.uint32(26).string(message.error)
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): ConnInfo_Knock {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseConnInfo_Knock()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break
          }

          message.knock = reader.bool()
          continue
        case 2:
          if (tag !== 16) {
            break
          }

          message.ack = reader.bool()
          continue
        case 3:
          if (tag !== 26) {
            break
          }

          message.error = reader.string()
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): ConnInfo_Knock {
    return {
      knock: isSet(object.knock) ? globalThis.Boolean(object.knock) : false,
      ack: isSet(object.ack) ? globalThis.Boolean(object.ack) : false,
      error: isSet(object.error) ? globalThis.String(object.error) : '',
    }
  },

  toJSON(message: ConnInfo_Knock): unknown {
    const obj: any = {}
    if (message.knock !== false) {
      obj.knock = message.knock
    }
    if (message.ack !== false) {
      obj.ack = message.ack
    }
    if (message.error !== '') {
      obj.error = message.error
    }
    return obj
  },

  create<I extends Exact<DeepPartial<ConnInfo_Knock>, I>>(base?: I): ConnInfo_Knock {
    return ConnInfo_Knock.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<ConnInfo_Knock>, I>>(
    object: I,
  ): ConnInfo_Knock {
    const message = createBaseConnInfo_Knock()
    message.knock = object.knock ?? false
    message.ack = object.ack ?? false
    message.error = object.error ?? ''
    return message
  },
}
// ...
export interface GRPCBrokerServer extends UntypedServiceImplementation {
  startStream: handleBidiStreamingCall<ConnInfo, ConnInfo>
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>

type KeysOfUnion<T> = T extends T ? keyof T : never
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never
    }